refactor(workflow): split recovery helpers
