+++ src/workflow/action_features.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Mapping
+from collections.abc import Callable, Mapping
 from dataclasses import dataclass
 from typing import Literal, cast
 
@@ -77,6 +77,33 @@ class ActionFeatureDerivation:
         }
 
 
+@dataclass(frozen=True)
+class ActionFeatureContext:
+    runtime_state: Mapping[str, object]
+    normalized_stage: str | None
+    normalized_failure_type: FailureType | None
+    retry_exhausted: bool
+    safety_blocked: bool
+    candidate: Mapping[str, object]
+    completed_step_count: int | None
+    failed_step_index: int | None
+    p_success: float
+    p_structural_failure: float
+    recovery_margin: float
+    expected_remaining_cost: float
+    evidence_sufficiency: float
+    has_context: bool
+
+
+FeatureBuilder = Callable[[ActionFeatureContext, dict[str, DerivedActionFeature]], DerivedActionFeature]
+
+
+@dataclass(frozen=True)
+class FeatureRule:
+    name: str
+    derive: FeatureBuilder
+
+
 def derive_action_features(
     *,
     runtime_state: Mapping[str, object],
@@ -88,315 +115,379 @@ def derive_action_features(
     completed_step_count: int | None = None,
     failed_step_index: int | None = None,
 ) -> ActionFeatureDerivation:
-    """从运行时状态和失败上下文派生动作效用特征。
+    """从运行时状态和失败上下文派生动作效用特征。"""
 
-    Args:
-        runtime_state: 当前运行时状态摘要。
-        stage_id: 失败或决策所在阶段。
-        failure_type: 失败类型。
-        retry_exhausted: 当前步骤是否已耗尽 retry。
-        safety_blocked: SafetyAgent 是否阻止继续。
-        candidate_summary: 候选或目标评分摘要。
-        completed_step_count: 已完成步骤数。
-        failed_step_index: 失败步骤索引。
-
-    Returns:
-        带 value/source/source_fields/reason 的派生结果。
-    """
-
-    normalized_failure_type = _normalize_failure_type(failure_type)
-    normalized_stage = _normalize_text(stage_id)
-    candidate = candidate_summary or {}
-
-    p_success = _unit_or_default(runtime_state.get("p_success"), 0.5)
-    p_structural_failure = _unit_or_default(
-        runtime_state.get("p_structural_failure"), 0.25
-    )
-    recovery_margin = _unit_or_default(runtime_state.get("recovery_margin"), 0.6)
-    expected_remaining_cost = _float_or_default(
-        runtime_state.get("expected_remaining_cost"), 1.0
-    )
-    evidence_sufficiency = _unit_or_default(
-        runtime_state.get("evidence_sufficiency"), 0.5
-    )
-    has_context = _has_contextual_signal(
+    context = _build_feature_context(
         runtime_state=runtime_state,
-        candidate_summary=candidate_summary,
-        stage_id=normalized_stage,
-        failure_type=normalized_failure_type,
+        stage_id=stage_id,
+        failure_type=failure_type,
         retry_exhausted=retry_exhausted,
         safety_blocked=safety_blocked,
+        candidate_summary=candidate_summary,
         completed_step_count=completed_step_count,
         failed_step_index=failed_step_index,
     )
-
     features: dict[str, DerivedActionFeature] = {}
+    for rule in FEATURE_RULES:
+        features[rule.name] = rule.derive(context, features)
+    values = {name: features[name].value for name in ACTION_FEATURE_NAMES}
+    return ActionFeatureDerivation(values=values, features=features)
 
+
+def _build_feature_context(
+    *,
+    runtime_state: Mapping[str, object],
+    stage_id: str | None,
+    failure_type: FailureType | str | None,
+    retry_exhausted: bool,
+    safety_blocked: bool,
+    candidate_summary: Mapping[str, object] | None,
+    completed_step_count: int | None,
+    failed_step_index: int | None,
+) -> ActionFeatureContext:
+    normalized_failure_type = _normalize_failure_type(failure_type)
+    normalized_stage = _normalize_text(stage_id)
+    candidate = candidate_summary or {}
+    return ActionFeatureContext(
+        runtime_state=runtime_state,
+        normalized_stage=normalized_stage,
+        normalized_failure_type=normalized_failure_type,
+        retry_exhausted=retry_exhausted,
+        safety_blocked=safety_blocked,
+        candidate=candidate,
+        completed_step_count=completed_step_count,
+        failed_step_index=failed_step_index,
+        p_success=_unit_or_default(runtime_state.get("p_success"), 0.5),
+        p_structural_failure=_unit_or_default(
+            runtime_state.get("p_structural_failure"), 0.25
+        ),
+        recovery_margin=_unit_or_default(runtime_state.get("recovery_margin"), 0.6),
+        expected_remaining_cost=_float_or_default(
+            runtime_state.get("expected_remaining_cost"), 1.0
+        ),
+        evidence_sufficiency=_unit_or_default(
+            runtime_state.get("evidence_sufficiency"), 0.5
+        ),
+        has_context=_has_contextual_signal(
+            runtime_state=runtime_state,
+            candidate_summary=candidate_summary,
+            stage_id=normalized_stage,
+            failure_type=normalized_failure_type,
+            retry_exhausted=retry_exhausted,
+            safety_blocked=safety_blocked,
+            completed_step_count=completed_step_count,
+            failed_step_index=failed_step_index,
+        ),
+    )
+
+
+def _derive_budget_pressure_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    _ = features
     budget_pressure = coerce_optional_budget_pressure(
-        runtime_state.get("budget_pressure")
+        context.runtime_state.get("budget_pressure")
     )
     if budget_pressure is not None:
-        features["budget_pressure"] = _feature(
+        return _feature(
             budget_pressure,
             "observed",
             ("runtime_state.budget_pressure",),
             "Runtime state provided budget pressure explicitly.",
             upper=1.5,
         )
-    else:
-        budget_derivation = derive_budget_pressure(
-            expected_remaining_cost=expected_remaining_cost,
-            budget_cap=coerce_optional_budget_cap(runtime_state.get("budget_cap")),
-        )
-        features["budget_pressure"] = _feature(
-            budget_derivation.budget_pressure,
-            "inferred" if "expected_remaining_cost" in runtime_state else "default",
-            budget_derivation.source_fields,
-            budget_derivation.reason,
-            upper=1.5,
-        )
+    budget_derivation = derive_budget_pressure(
+        expected_remaining_cost=context.expected_remaining_cost,
+        budget_cap=coerce_optional_budget_cap(context.runtime_state.get("budget_cap")),
+    )
+    return _feature(
+        budget_derivation.budget_pressure,
+        "inferred" if "expected_remaining_cost" in context.runtime_state else "default",
+        budget_derivation.source_fields,
+        budget_derivation.reason,
+        upper=1.5,
+    )
 
-    budget_pressure_value = features["budget_pressure"].value
 
-    local_patchability = _observed_or_none(runtime_state, "local_patchability")
-    if local_patchability is not None:
-        features["local_patchability"] = _feature(
-            local_patchability,
+def _derive_local_patchability_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    _ = features
+    observed = _observed_or_none(context.runtime_state, "local_patchability")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.local_patchability",),
             "Runtime state provided local patchability explicitly.",
         )
-    else:
-        patch_value = (
-            0.45 * recovery_margin
-            + 0.35 * (1.0 - p_structural_failure)
-            + 0.20 * evidence_sufficiency
-        )
-        if normalized_stage == "S1":
-            patch_value += 0.18
-        if normalized_failure_type in {FailureType.RETRYABLE, FailureType.TOOL_ERROR}:
-            patch_value += 0.10
-        if retry_exhausted and not safety_blocked:
-            patch_value += 0.05
-        if safety_blocked:
-            patch_value -= 0.25
-        if normalized_failure_type == FailureType.SAFETY_BLOCK:
-            patch_value -= 0.15
-        features["local_patchability"] = _feature(
-            patch_value,
-            "inferred" if has_context else "default",
-            (
-                "runtime_state.recovery_margin",
-                "runtime_state.p_structural_failure",
-                "runtime_state.evidence_sufficiency",
-                "failure_context.stage_id",
-                "failure_context.failure_type",
-                "failure_context.retry_exhausted",
-                "failure_context.safety_blocked",
-            ),
-            "Local patchability is inferred from recovery headroom, failure pressure, evidence, and failure locality.",
-        )
+    patch_value = (
+        0.45 * context.recovery_margin
+        + 0.35 * (1.0 - context.p_structural_failure)
+        + 0.20 * context.evidence_sufficiency
+    )
+    if context.normalized_stage == "S1":
+        patch_value += 0.18
+    if context.normalized_failure_type in {FailureType.RETRYABLE, FailureType.TOOL_ERROR}:
+        patch_value += 0.10
+    if context.retry_exhausted and not context.safety_blocked:
+        patch_value += 0.05
+    if context.safety_blocked:
+        patch_value -= 0.25
+    if context.normalized_failure_type == FailureType.SAFETY_BLOCK:
+        patch_value -= 0.15
+    return _feature(
+        patch_value,
+        "inferred" if context.has_context else "default",
+        (
+            "runtime_state.recovery_margin",
+            "runtime_state.p_structural_failure",
+            "runtime_state.evidence_sufficiency",
+            "failure_context.stage_id",
+            "failure_context.failure_type",
+            "failure_context.retry_exhausted",
+            "failure_context.safety_blocked",
+        ),
+        "Local patchability is inferred from recovery headroom, failure pressure, evidence, and failure locality.",
+    )
 
-    prefix_preservability = _observed_or_none(runtime_state, "prefix_preservability")
-    if prefix_preservability is not None:
-        features["prefix_preservability"] = _feature(
-            prefix_preservability,
+
+def _derive_prefix_preservability_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    observed = _observed_or_none(context.runtime_state, "prefix_preservability")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.prefix_preservability",),
             "Runtime state provided prefix preservability explicitly.",
         )
-    else:
-        prefix_value = (
-            0.50 * recovery_margin
-            + 0.30 * evidence_sufficiency
-            + 0.20 * (1.0 - min(budget_pressure_value, 1.0))
-        )
-        source_fields = [
-            "runtime_state.recovery_margin",
-            "runtime_state.evidence_sufficiency",
-            "derived.budget_pressure",
-        ]
-        if failed_step_index is not None and failed_step_index > 0:
-            prefix_value += 0.10
-            source_fields.append("failure_context.failed_step_index")
-        elif failed_step_index == 0:
-            prefix_value -= 0.15
-            source_fields.append("failure_context.failed_step_index")
-        if completed_step_count is not None and completed_step_count > 0:
-            prefix_value += 0.08
-            source_fields.append("failure_context.completed_step_count")
-        features["prefix_preservability"] = _feature(
-            prefix_value,
-            "inferred" if has_context else "default",
-            tuple(source_fields),
-            "Prefix preservability is inferred from recovery margin, evidence sufficiency, budget pressure, and known execution progress.",
-        )
+    prefix_value = (
+        0.50 * context.recovery_margin
+        + 0.30 * context.evidence_sufficiency
+        + 0.20 * (1.0 - min(features["budget_pressure"].value, 1.0))
+    )
+    source_fields = [
+        "runtime_state.recovery_margin",
+        "runtime_state.evidence_sufficiency",
+        "derived.budget_pressure",
+    ]
+    if context.failed_step_index is not None and context.failed_step_index > 0:
+        prefix_value += 0.10
+        source_fields.append("failure_context.failed_step_index")
+    elif context.failed_step_index == 0:
+        prefix_value -= 0.15
+        source_fields.append("failure_context.failed_step_index")
+    if context.completed_step_count is not None and context.completed_step_count > 0:
+        prefix_value += 0.08
+        source_fields.append("failure_context.completed_step_count")
+    return _feature(
+        prefix_value,
+        "inferred" if context.has_context else "default",
+        tuple(source_fields),
+        "Prefix preservability is inferred from recovery margin, evidence sufficiency, budget pressure, and known execution progress.",
+    )
 
-    evidence_reusability = _observed_or_none(runtime_state, "evidence_reusability")
-    if evidence_reusability is not None:
-        features["evidence_reusability"] = _feature(
-            evidence_reusability,
+
+def _derive_evidence_reusability_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    observed = _observed_or_none(context.runtime_state, "evidence_reusability")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.evidence_reusability",),
             "Runtime state provided evidence reusability explicitly.",
         )
-    else:
-        candidate_evidence = _nested_numeric(
-            candidate,
-            ("posterior_objective", "evidence_sufficiency"),
+    candidate_evidence = _nested_numeric(
+        context.candidate,
+        ("posterior_objective", "evidence_sufficiency"),
+    )
+    if candidate_evidence is not None:
+        return _feature(
+            candidate_evidence,
+            "observed",
+            ("candidate_summary.posterior_objective.evidence_sufficiency",),
+            "Candidate posterior objective provided reusable evidence sufficiency.",
         )
-        if candidate_evidence is not None:
-            features["evidence_reusability"] = _feature(
-                candidate_evidence,
-                "observed",
-                ("candidate_summary.posterior_objective.evidence_sufficiency",),
-                "Candidate posterior objective provided reusable evidence sufficiency.",
-            )
-        else:
-            features["evidence_reusability"] = _feature(
-                0.65 * evidence_sufficiency
-                + 0.35 * features["prefix_preservability"].value,
-                "inferred" if has_context else "default",
-                (
-                    "runtime_state.evidence_sufficiency",
-                    "derived.prefix_preservability",
-                ),
-                "Evidence reusability is inferred from evidence sufficiency and prefix preservability.",
-            )
+    return _feature(
+        0.65 * context.evidence_sufficiency
+        + 0.35 * features["prefix_preservability"].value,
+        "inferred" if context.has_context else "default",
+        (
+            "runtime_state.evidence_sufficiency",
+            "derived.prefix_preservability",
+        ),
+        "Evidence reusability is inferred from evidence sufficiency and prefix preservability.",
+    )
 
-    budget_relief = _observed_or_none(runtime_state, "budget_relief")
-    if budget_relief is not None:
-        features["budget_relief"] = _feature(
-            budget_relief,
+
+def _derive_budget_relief_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    observed = _observed_or_none(context.runtime_state, "budget_relief")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.budget_relief",),
             "Runtime state provided budget relief explicitly.",
         )
+    if context.expected_remaining_cost >= 1.0 and features["prefix_preservability"].value >= 0.55:
+        relief_value = 0.60
+    elif context.expected_remaining_cost >= 1.0:
+        relief_value = 0.45
     else:
-        if expected_remaining_cost >= 1.0 and features["prefix_preservability"].value >= 0.55:
-            relief_value = 0.60
-        elif expected_remaining_cost >= 1.0:
-            relief_value = 0.45
-        else:
-            relief_value = 0.35
-        features["budget_relief"] = _feature(
-            relief_value,
-            "inferred" if "expected_remaining_cost" in runtime_state else "default",
-            (
-                "runtime_state.expected_remaining_cost",
-                "derived.prefix_preservability",
-            ),
-            "Budget relief is conservative because suffix replan is not assumed to be free.",
-        )
+        relief_value = 0.35
+    return _feature(
+        relief_value,
+        "inferred" if "expected_remaining_cost" in context.runtime_state else "default",
+        (
+            "runtime_state.expected_remaining_cost",
+            "derived.prefix_preservability",
+        ),
+        "Budget relief is conservative because suffix replan is not assumed to be free.",
+    )
 
-    goal_realignment = _observed_or_none(runtime_state, "goal_realignment")
-    if goal_realignment is not None:
-        features["goal_realignment"] = _feature(
-            goal_realignment,
+
+def _derive_goal_realignment_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    _ = features
+    observed = _observed_or_none(context.runtime_state, "goal_realignment")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.goal_realignment",),
             "Runtime state provided goal realignment explicitly.",
         )
-    else:
-        realignment_value = (
-            0.40 * (1.0 - p_success)
-            + 0.30 * p_structural_failure
-            + 0.30 * (1.0 - evidence_sufficiency)
+    realignment_value = (
+        0.40 * (1.0 - context.p_success)
+        + 0.30 * context.p_structural_failure
+        + 0.30 * (1.0 - context.evidence_sufficiency)
+    )
+    objective_gap = _nested_numeric(context.candidate, ("objective_gap",))
+    objective_gap_source = "candidate_summary.objective_gap"
+    if objective_gap is None:
+        objective_gap = _nested_numeric(
+            context.candidate,
+            ("posterior_objective", "objective_gap"),
         )
-        objective_gap = _nested_numeric(candidate, ("objective_gap",))
-        objective_gap_source = "candidate_summary.objective_gap"
-        if objective_gap is None:
-            objective_gap = _nested_numeric(
-                candidate,
-                ("posterior_objective", "objective_gap"),
-            )
-            objective_gap_source = "candidate_summary.posterior_objective.objective_gap"
-        source_fields = [
-            "runtime_state.p_success",
-            "runtime_state.p_structural_failure",
-            "runtime_state.evidence_sufficiency",
-        ]
-        if objective_gap is not None:
-            realignment_value = max(realignment_value, objective_gap)
-            source_fields.append(objective_gap_source)
-        features["goal_realignment"] = _feature(
-            realignment_value,
-            "inferred" if has_context else "default",
-            tuple(source_fields),
-            "Goal realignment is inferred from low success, structural pressure, missing evidence, and optional objective gap.",
-        )
+        objective_gap_source = "candidate_summary.posterior_objective.objective_gap"
+    source_fields = [
+        "runtime_state.p_success",
+        "runtime_state.p_structural_failure",
+        "runtime_state.evidence_sufficiency",
+    ]
+    if objective_gap is not None:
+        realignment_value = max(realignment_value, objective_gap)
+        source_fields.append(objective_gap_source)
+    return _feature(
+        realignment_value,
+        "inferred" if context.has_context else "default",
+        tuple(source_fields),
+        "Goal realignment is inferred from low success, structural pressure, missing evidence, and optional objective gap.",
+    )
 
-    safety_terminality = _observed_or_none(runtime_state, "safety_terminality")
-    if safety_terminality is not None:
-        features["safety_terminality"] = _feature(
-            safety_terminality,
+
+def _derive_safety_terminality_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    _ = features
+    observed = _observed_or_none(context.runtime_state, "safety_terminality")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.safety_terminality",),
             "Runtime state provided safety terminality explicitly.",
         )
-    else:
-        safety_value = 1.0 if (
-            safety_blocked or normalized_failure_type == FailureType.SAFETY_BLOCK
-        ) else 0.0
-        features["safety_terminality"] = _feature(
-            safety_value,
-            "inferred" if safety_value > 0.0 else "default",
-            (
-                "failure_context.safety_blocked",
-                "failure_context.failure_type",
-            ),
-            "Safety terminality is only high when safety blocks continuation.",
-        )
+    safety_value = 1.0 if (
+        context.safety_blocked
+        or context.normalized_failure_type == FailureType.SAFETY_BLOCK
+    ) else 0.0
+    return _feature(
+        safety_value,
+        "inferred" if safety_value > 0.0 else "default",
+        (
+            "failure_context.safety_blocked",
+            "failure_context.failure_type",
+        ),
+        "Safety terminality is only high when safety blocks continuation.",
+    )
 
-    intervention_value = _observed_or_none(runtime_state, "intervention_value")
-    if intervention_value is not None:
-        features["intervention_value"] = _feature(
-            intervention_value,
+
+def _derive_intervention_value_feature(
+    context: ActionFeatureContext,
+    features: dict[str, DerivedActionFeature],
+) -> DerivedActionFeature:
+    observed = _observed_or_none(context.runtime_state, "intervention_value")
+    if observed is not None:
+        return _feature(
+            observed,
             "observed",
             ("runtime_state.intervention_value",),
             "Runtime state provided intervention value explicitly.",
         )
-    elif not has_context:
-        features["intervention_value"] = _feature(
+    if not context.has_context:
+        return _feature(
             0.5,
             "default",
             (),
             "No reliable runtime context is available, so intervention value defaults to neutral 0.5.",
         )
-    else:
-        uncertainty = _clip(
-            1.0 - abs(p_success - (1.0 - p_structural_failure))
-        )
-        manual_salvageability = _clip(
-            0.55 * features["local_patchability"].value
-            + 0.45 * features["prefix_preservability"].value
-        )
-        artifact_salience = _clip(
-            0.60 * evidence_sufficiency + 0.40 * recovery_margin
-        )
-        decision_gap = _clip(
-            0.50 + 0.25 * recovery_margin - 0.25 * min(budget_pressure_value, 1.0)
-        )
-        features["intervention_value"] = _feature(
-            0.30 * uncertainty
-            + 0.25 * manual_salvageability
-            + 0.25 * artifact_salience
-            + 0.20 * decision_gap,
-            "inferred",
-            (
-                "runtime_state.p_success",
-                "runtime_state.p_structural_failure",
-                "runtime_state.evidence_sufficiency",
-                "runtime_state.recovery_margin",
-                "derived.local_patchability",
-                "derived.prefix_preservability",
-                "derived.budget_pressure",
-            ),
-            "Intervention value is inferred from uncertainty, salvageability, artifact salience, and decision gap.",
-        )
+    budget_pressure_value = features["budget_pressure"].value
+    uncertainty = _clip(
+        1.0 - abs(context.p_success - (1.0 - context.p_structural_failure))
+    )
+    manual_salvageability = _clip(
+        0.55 * features["local_patchability"].value
+        + 0.45 * features["prefix_preservability"].value
+    )
+    artifact_salience = _clip(
+        0.60 * context.evidence_sufficiency + 0.40 * context.recovery_margin
+    )
+    decision_gap = _clip(
+        0.50 + 0.25 * context.recovery_margin - 0.25 * min(budget_pressure_value, 1.0)
+    )
+    return _feature(
+        0.30 * uncertainty
+        + 0.25 * manual_salvageability
+        + 0.25 * artifact_salience
+        + 0.20 * decision_gap,
+        "inferred",
+        (
+            "runtime_state.p_success",
+            "runtime_state.p_structural_failure",
+            "runtime_state.evidence_sufficiency",
+            "runtime_state.recovery_margin",
+            "derived.local_patchability",
+            "derived.prefix_preservability",
+            "derived.budget_pressure",
+        ),
+        "Intervention value is inferred from uncertainty, salvageability, artifact salience, and decision gap.",
+    )
 
-    values = {name: features[name].value for name in ACTION_FEATURE_NAMES}
-    return ActionFeatureDerivation(values=values, features=features)
+FEATURE_RULES: tuple[FeatureRule, ...] = (
+    FeatureRule("budget_pressure", _derive_budget_pressure_feature),
+    FeatureRule("local_patchability", _derive_local_patchability_feature),
+    FeatureRule("prefix_preservability", _derive_prefix_preservability_feature),
+    FeatureRule("evidence_reusability", _derive_evidence_reusability_feature),
+    FeatureRule("budget_relief", _derive_budget_relief_feature),
+    FeatureRule("goal_realignment", _derive_goal_realignment_feature),
+    FeatureRule("safety_terminality", _derive_safety_terminality_feature),
+    FeatureRule("intervention_value", _derive_intervention_value_feature),
+)
 
 
 def _feature(
