+++ src/workflow/plan_runner.py
@@ -1,16 +1,10 @@
 from __future__ import annotations
-import json
-from pathlib import Path
 from typing import Any, Protocol
 from src.agents.planner import PlannerAgent
 from src.infra.event_log_factory import make_candidate_validation_failed
 from src.models.contracts import (
-    PendingActionStatus,
-    PendingActionType,
     Plan,
-    ReplanRequest,
     StepResult,
-    now_iso,
 )
 from src.models.validation import (
     CandidateExecutionValidationError,
@@ -27,15 +21,11 @@ from src.storage.log_store import append_event, write_event_log
 from src.workflow.context import WorkflowContext
 from src.workflow.step_runner import StepRunner
 from src.workflow.patch_runner import PatchRunner, PendingPatch
-from src.workflow.pending_action import build_pending_action, enter_waiting_state
 from src.agents.safety import SafetyAgent
 from src.workflow.status import transition_task_status
 from src.workflow.snapshots import build_context_runtime_state_summary
 from src.workflow.recovery import (
-    WorkflowActionSelectorInput,
-    build_terminal_stop_candidate,
     resolve_workflow_action_route,
-    select_workflow_action,
 )
 from src.workflow.errors import (
     FailureType,
@@ -44,9 +34,15 @@ from src.workflow.errors import (
     classify_exception,
     is_retryable_failure,
 )
-from src.workflow.runtime_policy import resolve_runtime_policy, runtime_policy_trace
+from src.workflow.runtime_policy import runtime_policy_trace
+from src.workflow.plan_trace_builder import build_step_trace_data
+from src.workflow.plan_runner_recovery import (
+    PlanRunnerRecoveryMixin,
+    has_recovery_upgrade,
+    should_require_replan_confirm,
+)
 
-_AA_ALPHABET = set("ACDEFGHIKLMNPQRSTVWY")
+__all__ = ["PlanRunner", "_build_step_trace_data"]
 
 
 class StepRunnerLike(Protocol):
@@ -59,7 +55,7 @@ class StepRunnerLike(Protocol):
         """
 
 
-class PlanRunner:
+class PlanRunner(PlanRunnerRecoveryMixin):
     """PlanRunner: 顺序执行 Plan.steps, 并回写 WorkflowContext, 管理任务状态
 
     实现最小闭环的行为: 依次执行 Plan 中的步骤
@@ -149,6 +145,25 @@ class PlanRunner:
         )
 
     def run_plan(
+        self,
+        plan: Plan,
+        context: WorkflowContext,
+        *,
+        record: TaskRecord | None = None,
+        finalize_status: bool = True,
+        max_replans: int = 1,
+        resume_from_existing: bool = False,
+    ) -> Plan:
+        return self._run_plan_impl(
+            plan,
+            context,
+            record=record,
+            finalize_status=finalize_status,
+            max_replans=max_replans,
+            resume_from_existing=resume_from_existing,
+        )
+
+    def _run_plan_impl(
         self,
         plan: Plan,
         context: WorkflowContext,
@@ -287,7 +302,7 @@ class PlanRunner:
                         else None,
                     )
                     if step_result.status == "failed":
-                        if not _has_recovery_upgrade(step_result):
+                        if not has_recovery_upgrade(step_result):
                             blocked_by_safety = self._add_failed_step_safety_event(
                                 step_result,
                                 plan,
@@ -408,7 +423,7 @@ class PlanRunner:
             if (
                 context.status == InternalStatus.WAITING_REPLAN
                 and max_replans > 0
-                and not _should_require_replan_confirm(exc)
+                and not should_require_replan_confirm(exc)
             ):
                 replanned_plan = self._perform_replan(plan, context, record, exc)
                 if context.status == InternalStatus.WAITING_REPLAN:
@@ -423,7 +438,7 @@ class PlanRunner:
                 )
             if not (
                 context.status == InternalStatus.WAITING_REPLAN
-                and _should_require_replan_confirm(exc)
+                and should_require_replan_confirm(exc)
             ):
                 self._mark_failed(context, record, reason="plan_error")
             raise
@@ -487,427 +502,6 @@ class PlanRunner:
         else:
             context.step_results[result.step_id] = result
 
-    def _add_failed_step_safety_event(
-        self,
-        step_result: StepResult,
-        plan: Plan,
-        context: WorkflowContext,
-    ) -> bool:
-        step = next((s for s in plan.steps if s.id == step_result.step_id), None)
-        if step is None:
-            return False
-        safety_result = self._safety_agent.check_post_step(step, step_result, context)
-        self._add_safety_event(context, safety_result)
-        if safety_result.risk_flags:
-            step_result.risk_flags = safety_result.risk_flags
-        return safety_result.action == "block"
-
-    def _build_replan_explanation(
-        self,
-        reason: str,
-        failed_result: StepResult,
-    ) -> str:
-        failure_context = self._summarize_failure_result(failed_result)
-        options = self._build_replan_options(failure_context.get("failure_code"))
-        payload = {
-            "action_name": self._extract_workflow_action(failed_result),
-            "reason": reason,
-            "failure": failure_context,
-            "options": options,
-        }
-        return (
-            f"replan requested: {reason}; "
-            f"context={json.dumps(payload, ensure_ascii=True)}"
-        )
-
-    def _build_basic_replan_explanation(self, reason: str) -> str:
-        payload = {
-            "reason": reason,
-            "options": ["replan", "cancel"],
-        }
-        return (
-            f"replan requested: {reason}; "
-            f"context={json.dumps(payload, ensure_ascii=True)}"
-        )
-
-    def _build_replan_options(self, failure_code: object) -> list[str]:
-        options = ["replan", "cancel"]
-        if isinstance(failure_code, str) and failure_code.startswith("NIM_"):
-            options.append("switch_to_local_esmfold")
-        if isinstance(failure_code, str) and failure_code.startswith("S3_"):
-            options.append("suffix_replan")
-        return options
-
-    def _build_stop_explanation(self, failed_result: StepResult) -> str:
-        payload = {
-            "action_name": "stop",
-            "terminal_policy": "stop",
-            "failure": self._summarize_failure_result(failed_result),
-            "options": ["accept_terminal_stop", "continue", "cancel"],
-        }
-        return (
-            "adaptive terminal_stop requested; "
-            f"context={json.dumps(payload, ensure_ascii=True)}"
-        )
-
-    def _summarize_failure_result(self, result: StepResult) -> dict:
-        failure_code = None
-        failure_reason = result.error_message
-        if result.risk_flags:
-            failure_code = result.risk_flags[0].code
-            if result.risk_flags[0].message:
-                failure_reason = result.risk_flags[0].message
-        if not failure_code and isinstance(result.error_details, dict):
-            failure_code = result.error_details.get("failure_code")
-        patch_meta = result.metrics.get("patch")
-        recovery_meta = result.metrics.get("recovery")
-        return {
-            "step_id": result.step_id,
-            "tool": result.tool,
-            "failure_type": result.failure_type,
-            "failure_code": failure_code,
-            "failure_reason": failure_reason,
-            "risk_flags": [flag.model_dump() for flag in result.risk_flags],
-            "patch": patch_meta if isinstance(patch_meta, dict) else {},
-            "recovery": recovery_meta if isinstance(recovery_meta, dict) else {},
-        }
-
-    def _coerce_failure_type(self, value) -> FailureType:
-        """将 StepResult.failure_type 统一为 FailureType 枚举"""
-        if isinstance(value, FailureType):
-            return value
-        if isinstance(value, str):
-            try:
-                return FailureType(value)
-            except ValueError:
-                pass
-        return FailureType.NON_RETRYABLE
-
-    def _extract_workflow_action(self, result: StepResult) -> str | None:
-        action = result.metrics.get("workflow_action")
-        if isinstance(action, str) and action:
-            return action
-        return None
-
-    def _ensure_step_workflow_action(
-        self,
-        context: WorkflowContext,
-        step_result: StepResult,
-    ) -> None:
-        if self._extract_workflow_action(step_result):
-            return
-        runtime_state_summary = build_context_runtime_state_summary(
-            context,
-            require_runtime_state=True,
-        )
-        decision = select_workflow_action(
-            WorkflowActionSelectorInput(
-                phase="execution",
-                stage_id=(
-                    step_result.outputs.get("stage_id")
-                    if isinstance(step_result.outputs, dict)
-                    else None
-                ),
-                failure_code=(
-                    step_result.error_details.get("failure_code")
-                    if isinstance(step_result.error_details, dict)
-                    else None
-                ),
-                failure_type=step_result.failure_type,
-                retry_exhausted=bool(step_result.metrics.get("retry_exhausted")),
-                safety_blocked=step_result.failure_type == FailureType.SAFETY_BLOCK,
-                runtime_state_summary=runtime_state_summary,
-                runtime_policy=resolve_runtime_policy(context.task),
-            )
-        )
-        metrics = dict(step_result.metrics)
-        metrics["workflow_action"] = decision.action
-        metrics["workflow_action_mapped_flow"] = decision.mapped_flow
-        metrics["workflow_action_reason"] = decision.reason
-        metrics["workflow_action_evidence"] = dict(decision.evidence_source)
-        if (
-            step_result.status == "failed"
-            and decision.action != "continue"
-            and "retry_exhausted" not in metrics
-        ):
-            metrics["retry_exhausted"] = True
-        step_result.metrics = metrics
-
-    def _should_skip_step(self, step, context: WorkflowContext) -> bool:
-        """判断是否可跳过已成功且与当前计划一致的步骤"""
-        result = context.step_results.get(step.id)
-        if result is None:
-            return False
-        if result.status != "success":
-            return False
-        if result.tool != step.tool:
-            return False
-        return True
-
-    def _has_patch_applied(self, step_results: list[StepResult]) -> bool:
-        """判断当前批次结果中是否包含已应用的 patch"""
-        for result in step_results:
-            patch_meta = result.metrics.get("patch")
-            if isinstance(patch_meta, dict) and patch_meta.get("applied") is True:
-                return True
-        return False
-
-    def _request_replan(
-        self,
-        context: WorkflowContext,
-        record: TaskRecord | None,
-        *,
-        reason: str,
-        failure_type: FailureType,
-        message: str,
-        code: str | None = None,
-        step_id: str | None = None,
-        explanation: str | None = None,
-    ) -> None:
-        """触发 WAITING_REPLAN，并抛出 PlanRunError 交给上层处理"""
-        if context.status != InternalStatus.WAITING_REPLAN:
-            pending_action = build_pending_action(
-                task_id=context.task.task_id,
-                action_type=PendingActionType.REPLAN_CONFIRM,
-                candidates=[],
-                default_suggestion=None,
-                explanation=explanation or f"replan requested: {reason}",
-            )
-            enter_waiting_state(
-                context,
-                record,
-                pending_action,
-                InternalStatus.WAITING_REPLAN,
-                reason=reason,
-            )
-            transition_task_status(
-                context,
-                record,
-                InternalStatus.WAITING_REPLAN,
-                reason=reason,
-            )
-        raise PlanRunError(
-            failure_type=failure_type,
-            message=message,
-            step_id=step_id,
-            code=code,
-        )
-
-    def _request_stop(
-        self,
-        context: WorkflowContext,
-        record: TaskRecord | None,
-        *,
-        failure_type: FailureType,
-        message: str,
-        step_id: str | None = None,
-        explanation: str | None = None,
-        failed_result: StepResult | None = None,
-    ) -> None:
-        if context.plan is None:
-            raise PlanRunError(
-                failure_type=failure_type,
-                message=message,
-                step_id=step_id,
-                code="ADAPTIVE_STOP_REQUESTED",
-            )
-        failure_code = None
-        failure_reason = message
-        if failed_result is not None:
-            failure_summary = self._summarize_failure_result(failed_result)
-            raw_code = failure_summary.get("failure_code")
-            if isinstance(raw_code, str) and raw_code:
-                failure_code = raw_code
-            raw_reason = failure_summary.get("failure_reason")
-            if isinstance(raw_reason, str) and raw_reason:
-                failure_reason = raw_reason
-        runtime_state_summary = build_context_runtime_state_summary(
-            context,
-            require_runtime_state=True,
-        )
-        terminal_stop_candidate = build_terminal_stop_candidate(
-            plan=context.plan,
-            step_id=step_id,
-            failure_type=failure_type,
-            failure_code=failure_code,
-            failure_reason=failure_reason,
-            runtime_state_summary=runtime_state_summary,
-            explanation=explanation,
-        )
-        pending_action = build_pending_action(
-            task_id=context.task.task_id,
-            action_type=PendingActionType.REPLAN_CONFIRM,
-            candidates=[terminal_stop_candidate],
-            default_suggestion=terminal_stop_candidate.candidate_id,
-            default_recommendation=terminal_stop_candidate.candidate_id,
-            explanation=explanation or "adaptive terminal_stop requested",
-            metadata={
-                "workflow_action": "stop",
-                "workflow_action_mapped_flow": "stop",
-                "workflow_action_reason": failure_reason,
-                "workflow_action_target": "failed",
-                "terminal_policy": "stop",
-            },
-        )
-        enter_waiting_state(
-            context,
-            record,
-            pending_action,
-            InternalStatus.WAITING_REPLAN,
-            reason="terminal_stop_requested",
-        )
-        transition_task_status(
-            context,
-            record,
-            InternalStatus.WAITING_REPLAN,
-            reason="terminal_stop_requested",
-        )
-        raise PlanRunError(
-            failure_type=failure_type,
-            message=message,
-            step_id=step_id,
-            code="ADAPTIVE_STOP_REQUESTED",
-        )
-
-    def _perform_replan(
-        self,
-        plan: Plan,
-        context: WorkflowContext,
-        record: TaskRecord | None,
-        error: PlanRunError,
-    ) -> Plan:
-        """执行最小再规划闭环：WAITING_REPLAN → REPLANNING → RUNNING"""
-        transition_task_status(
-            context,
-            record,
-            InternalStatus.REPLANNING,
-            reason="replan_requested",
-        )
-        self._resolve_pending_replan_action(context, record)
-        request = ReplanRequest(
-            task_id=context.task.task_id,
-            original_plan=plan,
-            failed_steps=[error.step_id] if error.step_id else [],
-            safety_events=list(context.safety_events),
-            reason=str(error),
-        )
-        try:
-            try:
-                replan_top_k = self._planner.replan_top_k(
-                    request,
-                    k=_resolve_top_k(context.task.constraints.get("replan_top_k"), default=3),
-                    runtime_state=context.runtime_state,
-                )
-                gate = self._planner.evaluate_top_k_gate(
-                    candidate_kind="replan",
-                    top_k_result=replan_top_k,
-                    task_constraints=context.task.constraints,
-                )
-                if gate.requires_hitl:
-                    waiting_action = build_pending_action(
-                        task_id=context.task.task_id,
-                        action_type=PendingActionType.REPLAN_CONFIRM,
-                        candidates=replan_top_k.candidates,
-                        default_suggestion=replan_top_k.default_recommendation,
-                        default_recommendation=replan_top_k.default_recommendation,
-                        explanation=f"{replan_top_k.explanation} gate={gate.reason}",
-                    )
-                    enter_waiting_state(
-                        context,
-                        record,
-                        waiting_action,
-                        InternalStatus.WAITING_REPLAN,
-                        reason=gate.reason,
-                    )
-                    transition_task_status(
-                        context,
-                        record,
-                        InternalStatus.WAITING_REPLAN,
-                        reason=gate.reason,
-                    )
-                    return plan
-
-                ordered_candidates = _select_replan_candidates(
-                    replan_top_k.candidates,
-                    default_recommendation=replan_top_k.default_recommendation,
-                    preferred_action=(
-                        "suffix_replan"
-                        if error.code == "SUFFIX_REPLAN_REQUESTED"
-                        else None
-                    ),
-                )
-                selected = ordered_candidates[0] if ordered_candidates else None
-                if selected is None:
-                    raise ValueError("replan_top_k returned empty candidates")
-                payload = selected.structured_payload
-                if not isinstance(payload, Plan):
-                    raise ValueError("replan_top_k selected payload is not Plan")
-                replanned_plan = payload
-            except Exception:
-                # 回退到旧 replan()，保持既有 Planner stub 兼容
-                replanned_plan = self._planner.replan(request)
-        except Exception as exc:
-            transition_task_status(
-                context,
-                record,
-                InternalStatus.FAILED,
-                reason="replan_failed",
-            )
-            raise PlanRunError(
-                failure_type=classify_exception(exc),
-                message=f"Replan failed: {exc}",
-                step_id=error.step_id,
-                code="REPLAN_FAILED",
-                cause=exc,
-            ) from exc
-
-        if replanned_plan.task_id != context.task.task_id:
-            transition_task_status(
-                context,
-                record,
-                InternalStatus.FAILED,
-                reason="replan_failed",
-            )
-            raise PlanRunError(
-                failure_type=FailureType.NON_RETRYABLE,
-                message=(
-                    "Replan produced plan with mismatched task_id: "
-                    f"{replanned_plan.task_id} != {context.task.task_id}"
-                ),
-                step_id=error.step_id,
-                code="REPLAN_TASK_MISMATCH",
-            )
-
-        context.plan = replanned_plan
-        if record is not None:
-            record.plan = replanned_plan
-
-        transition_task_status(
-            context,
-            record,
-            InternalStatus.RUNNING,
-            reason="replan_succeeded",
-        )
-        return replanned_plan
-
-    def _resolve_pending_replan_action(
-        self,
-        context: WorkflowContext,
-        record: TaskRecord | None,
-    ) -> None:
-        action = context.pending_action
-        if (
-            action is None
-            or action.action_type != PendingActionType.REPLAN_CONFIRM
-            or action.status != PendingActionStatus.PENDING
-        ):
-            return
-        action.status = PendingActionStatus.CANCELLED
-        action.decided_at = now_iso()
-        context.pending_action = action
-        if record is not None:
-            record.pending_action = action
-
     def _mark_failed(
         self,
         context: WorkflowContext,
@@ -943,7 +537,7 @@ class PlanRunner:
             "state": context.status.value,
             "external_status": to_external_status(context.status).value,
         }
-        event_data = _build_step_trace_data(step_result)
+        event_data = build_step_trace_data(step_result)
         event_data.update(runtime_policy_trace(context.task))
         runtime_state_summary = build_context_runtime_state_summary(context)
         if runtime_state_summary is not None:
@@ -956,278 +550,5 @@ class PlanRunner:
         )
 
 
-def _should_require_replan_confirm(error: PlanRunError) -> bool:
-    """SafetyAgent blocks must wait for HITL replan confirmation."""
-    return (
-        error.failure_type == FailureType.SAFETY_BLOCK
-        or error.code in {
-            "SAFETY_TASK_INPUT_BLOCK",
-            "SAFETY_FINAL_BLOCK",
-            "SAFETY_POST_BLOCK",
-            "ADAPTIVE_STOP_REQUESTED",
-        }
-    )
-
-
-def _has_recovery_upgrade(step_result: StepResult) -> bool:
-    recovery = step_result.metrics.get("recovery")
-    if not isinstance(recovery, dict):
-        return False
-    upgrade_reason = recovery.get("upgrade_reason")
-    return isinstance(upgrade_reason, str) and bool(upgrade_reason)
-
-
 def _build_step_trace_data(step_result: StepResult) -> dict[str, Any]:
-    data: dict[str, Any] = {}
-    for field_name in (
-        "tool_id",
-        "adapter_id",
-        "execution_mode",
-        "provider",
-        "endpoint_type",
-        "remote_job_id",
-    ):
-        value = getattr(step_result, field_name, None)
-        if isinstance(value, str) and value:
-            data[field_name] = value
-    if isinstance(step_result.error_details, dict):
-        failure_code = step_result.error_details.get("failure_code")
-        if isinstance(failure_code, str) and failure_code:
-            data["failure_code"] = failure_code
-        remote_job_id = step_result.error_details.get("remote_job_id")
-        if isinstance(remote_job_id, str) and remote_job_id and "remote_job_id" not in data:
-            data["remote_job_id"] = remote_job_id
-
-    outputs = step_result.outputs if isinstance(step_result.outputs, dict) else {}
-    stage_id = outputs.get("stage_id")
-    if isinstance(stage_id, str) and stage_id:
-        data["stage_id"] = stage_id
-
-    patch_meta = step_result.metrics.get("patch")
-    if isinstance(patch_meta, dict) and patch_meta:
-        data["patch"] = {
-            "layer": patch_meta.get("layer"),
-            "from_tool": patch_meta.get("from_tool"),
-            "to_tool": patch_meta.get("to_tool"),
-            "capability_id": patch_meta.get("capability_id"),
-            "io_type": patch_meta.get("io_type"),
-            "adapter_mode": patch_meta.get("adapter_mode"),
-            "reason": patch_meta.get("reason"),
-            "ops": patch_meta.get("ops"),
-            "patched_status": patch_meta.get("patched_status"),
-        }
-
-    recovery_meta = step_result.metrics.get("recovery")
-    if isinstance(recovery_meta, dict) and recovery_meta:
-        data["recovery"] = {
-            "layer": recovery_meta.get("recovery_layer"),
-            "from_tool": recovery_meta.get("from_tool"),
-            "to_tool": recovery_meta.get("to_tool"),
-            "capability_id": recovery_meta.get("capability_id"),
-            "io_type": recovery_meta.get("io_type"),
-            "adapter_mode": recovery_meta.get("adapter_mode"),
-            "candidate_id": recovery_meta.get("candidate_id"),
-            "reason": recovery_meta.get("reason"),
-            "upgrade_reason": recovery_meta.get("upgrade_reason"),
-        }
-
-    fallback_meta = step_result.metrics.get("fallback")
-    if isinstance(fallback_meta, dict) and fallback_meta:
-        data["fallback"] = {
-            "fallback_kind": fallback_meta.get("fallback_kind"),
-            "reason": fallback_meta.get("reason"),
-            "from_tool_id": fallback_meta.get("from_tool_id"),
-            "to_tool_id": fallback_meta.get("to_tool_id"),
-            "from_execution_mode": fallback_meta.get("from_execution_mode"),
-            "to_execution_mode": fallback_meta.get("to_execution_mode"),
-            "from_adapter_id": fallback_meta.get("from_adapter_id"),
-            "to_adapter_id": fallback_meta.get("to_adapter_id"),
-            "capability_preserved": fallback_meta.get("capability_preserved"),
-        }
-
-    workflow_action = step_result.metrics.get("workflow_action")
-    if isinstance(workflow_action, str) and workflow_action:
-        data["action_name"] = workflow_action
-        data["workflow_action_reason"] = step_result.metrics.get(
-            "workflow_action_reason"
-        )
-        data["evidence_source"] = step_result.metrics.get(
-            "workflow_action_evidence"
-        )
-
-    s6_action = step_result.metrics.get("s6_recovery_action")
-    if isinstance(s6_action, str) and s6_action:
-        if "action_name" not in data:
-            data["action_name"] = s6_action
-        data["s6"] = {
-            "action": s6_action,
-            "trigger_stage_id": step_result.metrics.get("s6_trigger_stage_id"),
-            "trigger_failure_code": step_result.metrics.get("s6_trigger_failure_code"),
-        }
-
-    if stage_id == "S3":
-        input_summary = _summarize_quality_gate_inputs(step_result.inputs)
-        if input_summary:
-            data["input_summary"] = input_summary
-        reject_counts = outputs.get("reject_code_counts")
-        failed_rows = outputs.get("failed_samples")
-        failed_samples: list[dict[str, Any]] = []
-        if isinstance(failed_rows, list):
-            for item in failed_rows:
-                if not isinstance(item, dict):
-                    continue
-                failed_samples.append(
-                    {
-                        "candidate_id": item.get("candidate_id"),
-                        "reject_codes": item.get("reject_codes"),
-                        "reason": item.get("reason"),
-                    }
-                )
-        data["quality_gate"] = {
-            "pass_count": outputs.get("pass_count"),
-            "fail_count": outputs.get("fail_count"),
-            "pass_fail": outputs.get("pass_fail"),
-            "reject_code_counts": reject_counts if isinstance(reject_counts, dict) else {},
-            "failed_samples": failed_samples,
-        }
-    if outputs.get("capability_id") == "objective_scoring":
-        data["objective_scoring"] = {
-            "objective_score": outputs.get("objective_score"),
-            "objective_gap": step_result.metrics.get("objective_gap"),
-            "objective_progress": step_result.metrics.get("objective_progress"),
-            "default_recommendation": outputs.get("default_recommendation"),
-            "rank_reason": outputs.get("rank_reason"),
-            "warning_count": step_result.metrics.get("warning_count"),
-        }
-    return data
-
-
-def _summarize_quality_gate_inputs(inputs: Any) -> dict[str, Any]:
-    if not isinstance(inputs, dict):
-        return {}
-
-    summary: dict[str, Any] = {}
-    if "sequence" in inputs:
-        summary["sequence"] = _summarize_sequence_input(inputs.get("sequence"))
-    if "pdb_path" in inputs:
-        summary["pdb_path"] = _summarize_path_input(inputs.get("pdb_path"))
-
-    structure_results = inputs.get("structure_results")
-    if isinstance(structure_results, list):
-        summary["structure_results"] = {
-            "type": "list",
-            "count": len(structure_results),
-        }
-        first_item = next(
-            (item for item in structure_results if isinstance(item, dict)),
-            None,
-        )
-        if first_item is not None:
-            summary["structure_results"]["first_candidate"] = {
-                "candidate_id": first_item.get("candidate_id"),
-                "sequence": _summarize_sequence_input(first_item.get("sequence")),
-                "pdb_path": _summarize_path_input(first_item.get("pdb_path")),
-            }
-    return summary
-
-
-def _summarize_sequence_input(value: Any) -> dict[str, Any]:
-    summary: dict[str, Any] = {
-        "type": type(value).__name__,
-    }
-    if not isinstance(value, str):
-        if isinstance(value, list):
-            summary["count"] = len(value)
-        return summary
-
-    trimmed = value.strip()
-    uppercase = trimmed.upper()
-    invalid_chars = [
-        char
-        for char in uppercase
-        if char and char not in _AA_ALPHABET
-    ]
-    summary.update(
-        {
-            "length": len(value),
-            "preview": value[:48],
-            "symbolic_reference_like": _looks_like_symbolic_reference(value),
-            "valid_aa_chars": bool(trimmed) and not invalid_chars,
-            "invalid_char_count": len(invalid_chars),
-        }
-    )
-    return summary
-
-
-def _summarize_path_input(value: Any) -> dict[str, Any]:
-    summary: dict[str, Any] = {
-        "type": type(value).__name__,
-    }
-    if not isinstance(value, str):
-        return summary
-
-    symbolic_reference_like = _looks_like_symbolic_reference(value)
-    summary.update(
-        {
-            "value": value,
-            "symbolic_reference_like": symbolic_reference_like,
-            "exists": False if symbolic_reference_like else Path(value).exists(),
-        }
-    )
-    return summary
-
-
-def _looks_like_symbolic_reference(value: str) -> bool:
-    if "/" in value or "\\" in value:
-        return False
-    dot_count = value.count(".")
-    if dot_count >= 2:
-        return True
-    if dot_count == 1:
-        head = value.split(".", 1)[0]
-        return head.startswith("S")
-    return False
-
-
-def _resolve_top_k(value: object, *, default: int) -> int:
-    try:
-        parsed = int(value)
-    except (TypeError, ValueError):
-        return default
-    if parsed <= 0:
-        return 1
-    return parsed
-
-
-def _select_replan_candidates(
-    candidates: list[Any],
-    *,
-    default_recommendation: str | None,
-    preferred_action: str | None,
-) -> list[Any]:
-    preferred: list[tuple[int, str, Any]] = []
-    fallback: list[tuple[int, str, Any]] = []
-    for candidate in candidates:
-        metadata = candidate.metadata if isinstance(candidate.metadata, dict) else {}
-        shadow_action = metadata.get("shadow_action")
-        payload = candidate.structured_payload
-        replan_mode = None
-        if isinstance(payload, Plan):
-            payload_meta = payload.metadata if isinstance(payload.metadata, dict) else {}
-            raw_mode = payload_meta.get("replan_mode")
-            if isinstance(raw_mode, str):
-                replan_mode = raw_mode
-        rank = (
-            0
-            if default_recommendation is not None
-            and candidate.candidate_id == default_recommendation
-            else 1
-        )
-        row = (rank, candidate.candidate_id, candidate)
-        if preferred_action == "suffix_replan" and (
-            shadow_action == "suffix_replan" or replan_mode == "suffix_replan"
-        ):
-            preferred.append(row)
-        else:
-            fallback.append(row)
-    return [item[2] for item in sorted(preferred) + sorted(fallback)]
+    return build_step_trace_data(step_result)
